requirements.txt
werkzeug==2.2.3
flask==2.3.3
protobuf==4.24.0
grpcio==1.57.0
grpcio-tools==1.57.0
flask_cors==5.0.0
